Changing pages to ssr part 2: admin/timetable [nodeploy]
