+++ src/app/admin/users/page.tsx
@@ -1,9 +1,11 @@
-"use client";
-
 import CreateUsers from "components/CreateUsers";
 import OnlyRoles from "components/OnlyRoles";
 import PageWithHeader from "components/PageWithHeader";
 import React from "react";
+
+export const metadata = {
+    title: "MyBPHS - Admin / Felhasználók",
+};
 
 function Users() {
     return (
